chore: simplify in-construction identification
Use the `id` of the polycon to identify that a polycon is being created to avoid double init.

I don't think this is the most elegant solution but perhaps slightly less hacky.

I was also thinking to consider plugging in some market on `scope` as an alternative but haven't had a chance to test this.